Enable email sending configuration settings
The update enhances the code with a feature that allows or blocks email sending based on an environment variable 'SEND_EMAILS'. A new utility function 'isEmailingEnabled()' was added to the SiteUtil class to validate the existence of the 'SEND_EMAILS' environment variable. This is then utilized in EmailManager to determine whether to send out emails or not.